Title: 씨 오 더블유, Time: 152ms, Memory: 11MB, Status: Wrong Answer - Codetree

+++ 250525/씨 오 더블유/c-o-w.java
@@ -20,7 +20,7 @@ public class Main {
         System.out.print(dp());
     }
 
-    public static int dp() {
+    public static long dp() {
         for (int i = 1; i<=n; i++) {
             char a = alpha[i];
 
@@ -41,6 +41,6 @@ public class Main {
             }
         }
 
-        return nums[n][2];
+        return (long)nums[n][2];
     }
 }
